refactor: remove dependency on `/runtime/internal/*` subpaths
