+++ src/core/config/resolvers/open-api.ts
@@ -1,18 +1,20 @@
 import type { NitroOptions } from "nitropack/types";
+import { runtimeDir } from "nitropack/runtime/meta";
+import { join } from "pathe";
 
 export async function resolveOpenAPIOptions(options: NitroOptions) {
   if (options.dev && options.experimental.openAPI) {
     options.handlers.push({
       route: "/_nitro/openapi.json",
-      handler: "nitropack/runtime/internal/routes/openapi",
+      handler: join(runtimeDir, "internal/routes/openapi"),
     });
     options.handlers.push({
       route: "/_nitro/scalar",
-      handler: "nitropack/runtime/internal/routes/scalar",
+      handler: join(runtimeDir, "internal/routes/scalar"),
     });
     options.handlers.push({
       route: "/_nitro/swagger",
-      handler: "nitropack/runtime/internal/routes/swagger",
+      handler: join(runtimeDir, "internal/routes/swagger"),
     });
   }
 }
