+++ src/presets/_nitro/runtime/nitro-prerenderer.ts
@@ -1,6 +1,6 @@
 import "#nitro-internal-pollyfills";
 import { useNitroApp } from "nitropack/runtime";
-import { trapUnhandledNodeErrors } from "nitropack/runtime/internal/utils";
+import { trapUnhandledNodeErrors } from "nitropack/runtime/internal";
 
 const nitroApp = useNitroApp();
