+++ src/presets/bun/runtime/bun.ts
@@ -1,6 +1,6 @@
 import "#nitro-internal-pollyfills";
 import { useNitroApp } from "nitropack/runtime";
-import { startScheduleRunner } from "nitropack/runtime/internal/task";
+import { startScheduleRunner } from "nitropack/runtime/internal";
 
 import type {} from "bun";
 import wsAdapter from "crossws/adapters/bun";
