+++ src/presets/cloudflare/runtime/cloudflare-module.ts
@@ -1,8 +1,6 @@
 import "#nitro-internal-pollyfills";
-import { requestHasBody } from "nitropack/runtime/internal/utils";
-import { useNitroApp } from "nitropack/runtime";
-import { useRuntimeConfig } from "nitropack/runtime";
-import { runCronTasks } from "nitropack/runtime/internal/task";
+import { requestHasBody, runCronTasks } from "nitropack/runtime/internal";
+import { useNitroApp, useRuntimeConfig } from "nitropack/runtime";
 import { getPublicAssetMeta } from "#nitro-internal-virtual/public-assets";
 
 import { withoutBase } from "ufo";
