+++ src/presets/cloudflare/runtime/cloudflare-worker.ts
@@ -1,7 +1,6 @@
 import "#nitro-internal-pollyfills";
-import { requestHasBody } from "nitropack/runtime/internal/utils";
-import { useNitroApp } from "nitropack/runtime";
-import { useRuntimeConfig } from "nitropack/runtime";
+import { requestHasBody } from "nitropack/runtime/internal";
+import { useNitroApp, useRuntimeConfig } from "nitropack/runtime";
 import { getPublicAssetMeta } from "#nitro-internal-virtual/public-assets";
 
 import { withoutBase } from "ufo";
