+++ src/presets/deno/runtime/deno-deploy.ts
@@ -1,4 +1,5 @@
 import "#nitro-internal-pollyfills";
+import "./_deno-env-polyfill";
 import { useNitroApp } from "nitropack/runtime";
 
 import wsAdapter from "crossws/adapters/deno";
