+++ src/presets/deno/runtime/deno-server.ts
@@ -1,7 +1,8 @@
 import "#nitro-internal-pollyfills";
+import "./_deno-env-polyfill";
 import { useNitroApp } from "nitropack/runtime";
 import { useRuntimeConfig } from "nitropack/runtime";
-import { startScheduleRunner } from "nitropack/runtime/internal/task";
+import { startScheduleRunner } from "nitropack/runtime/internal";
 
 import destr from "destr";
 import wsAdapter from "crossws/adapters/deno";
