+++ src/presets/netlify/legacy/runtime/_deno-env-polyfill.ts
@@ -1,0 +1,1 @@
+Object.assign(process.env, Deno.env.toObject());
