+++ src/presets/netlify/legacy/runtime/netlify-builder.ts
@@ -1,4 +1,5 @@
 import "#nitro-internal-pollyfills";
+import "./_deno-env-polyfill";
 
 import { builder } from "@netlify/functions";
 import { lambda } from "./netlify-lambda";
