+++ src/presets/winterjs/runtime/winterjs.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck TODO: Remove after removing polyfills
 import "#nitro-internal-pollyfills";
-import { toBuffer } from "nitropack/runtime/internal/utils";
+import { toBuffer } from "nitropack/runtime/internal";
 import { useNitroApp } from "nitropack/runtime";
 import { toPlainHandler } from "h3";
 import { hasProtocol, joinURL } from "ufo";
