+++ src/runtime/internal/index.ts
@@ -1,0 +1,20 @@
+// Limited INTERNAL exports used by the presets runtime and nitro defaults
+
+export {
+  trapUnhandledNodeErrors,
+  normalizeCookieHeader,
+  requestHasBody,
+  joinHeaders,
+  toBuffer,
+} from "./utils";
+
+export {
+  normalizeLambdaIncomingHeaders,
+  normalizeLambdaOutgoingHeaders,
+  normalizeLambdaOutgoingBody,
+} from "./utils.lambda";
+
+export { startScheduleRunner, runCronTasks } from "./task";
+export { getAzureParsedCookiesFromHeaders } from "./utils.azure";
+export { getGracefulShutdownConfig, setupGracefulShutdown } from "./shutdown";
+export { getRouteRulesForPath } from "./route-rules";
